pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "onetoken"
version = "0.1.0"
description = "A library for running local LLM classification tasks on data"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Maxime Rivest", email = "mrive052@gmail.com"}
]
dependencies = [
    "pandas>=1.3.0",
    "transformers>=4.30.0",
    "llama-cpp-python>=0.2.0",
]

[project.optional-dependencies]
gpu = ["vllm>=0.2.0"]

[project.urls]
"Homepage" = "https://github.com/maximerivest/onetoken"
"Bug Tracker" = "https://github.com/maximerivest/onetoken/issues"

[tool.setuptools]
packages = ["onetoken"] 